requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.0.5=pyhd3eb1b0_0
bidict=0.21.2=pyhd3eb1b0_0
bzip2=1.0.8=h2bbff1b_5
ca-certificates=2024.3.11=haa95532_0
click=8.1.7=py310haa95532_0
colorama=0.4.6=py310haa95532_0
comm=0.2.1=py310haa95532_0
debugpy=1.6.7=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
exceptiongroup=1.2.0=py310haa95532_0
executing=0.8.3=pyhd3eb1b0_0
flask=2.2.5=py310haa95532_0
flask-socketio=5.3.1=py310haa95532_0
ipykernel=6.28.0=py310haa95532_0
ipython=8.20.0=py310haa95532_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jedi=0.18.1=py310haa95532_1
jinja2=3.1.3=py310haa95532_0
jupyter_client=8.6.0=py310haa95532_0
jupyter_core=5.5.0=py310haa95532_0
libffi=3.4.4=hd77b12b_0
libsodium=1.0.18=h62dcd97_0
markupsafe=2.1.3=py310h2bbff1b_0
matplotlib-inline=0.1.6=py310haa95532_0
nest-asyncio=1.6.0=py310haa95532_0
openssl=3.0.13=h2bbff1b_0
packaging=23.2=py310haa95532_0
parso=0.8.3=pyhd3eb1b0_0
pip=23.3.1=py310haa95532_0
platformdirs=3.10.0=py310haa95532_0
playsound=1.3.0=pyhd8ed1ab_0
prompt-toolkit=3.0.43=py310haa95532_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pygments=2.15.1=py310haa95532_1
python=3.10.14=he1021f5_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-engineio=4.1.0=pyhd3eb1b0_0
python-socketio=5.3.0=pyhd3eb1b0_0
pywin32=305=py310h2bbff1b_0
pyzmq=25.1.2=py310hd77b12b_0
setuptools=68.2.2=py310haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tk=8.6.12=h2bbff1b_0
tornado=6.3.3=py310h2bbff1b_0
traitlets=5.7.1=py310haa95532_0
tzdata=2024a=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.3.8=py310haa95532_0
wheel=0.41.2=py310haa95532_0
xz=5.4.6=h8cc25b3_0
zeromq=4.3.5=hd77b12b_0
zlib=1.2.13=h8cc25b3_0
gpsd-py3
